+++ src/bc/globals.d.ts
@@ -1,4 +1,6 @@
 declare global {
+  var unsafeWindow: Window & typeof globalThis;
+
   interface BCPlayer {
     ID?: number;
     MemberNumber: number;